requirements.txt
pyjwt
Flask
Flask-SQLAlchemy
Flask-WTF
passlib
psycopg2-binary